+++ src/tool/websearch.eml.ml
@@ -11,7 +11,7 @@ let show_form ~from ?(message="") ?output request =
     }
     </script>
 
-    <form method="POST" action="/" id="form">
+    <form method="POST" id="form">
       <%s! Dream.csrf_tag request %>
       <p>
       <input type="search" required="true" size="100"
@@ -40,7 +40,7 @@ let show_form ~from ?(message="") ?output request =
   </body>
   </html>
 
-let start ~header ss ~port ~dbpath () =
+let start ~header ss ~port ~dbpath ~path_in_url () =
   (*Common.Logger.set_debug true "e" ;*)
   let interface = "0.0.0.0" in
   Dream.run ~port ~interface
@@ -48,20 +48,24 @@ let start ~header ss ~port ~dbpath () =
   @@ Dream.memory_sessions
   @@ Dream.router [
 
-    Dream.get  "/"
+    Dream.get  ("/" ^ path_in_url)
       (fun request ->
         Dream.html (header ^ show_form ~from:0 request));
 
-    Dream.post "/"
+    Dream.post ("/" ^ path_in_url)
       (fun request ->
         match%lwt Dream.form request with
         | `Ok [ "from", from; "message", message; "search", _search ] ->
           Dream.log "from1 = %s" from ;
           let from = int_of_string from in (* XXX CSC exception XXX *)
           Dream.log "from2 = %d" from ;
+          if Timed.(!Common.Console.verbose) >= 2 then
+            Dream.log "Received request = %s" message;
           let output =
             Indexing.search_cmd_html ss ~from ~how_many:100
             message ~dbpath in
+          if Timed.(!Common.Console.verbose) >= 3 then
+            Dream.log "sending response: %s" output;
           Dream.html (show_form ~from ~message ~output request)
           (*Dream.stream (show_form_stream ~message ~output request)*)
         | _ ->