+++ src/tool/websearch.mli
@@ -1,2 +1,2 @@
 val start : header:string -> Core.Sig_state.sig_state
-    -> port:int -> dbpath:string -> unit -> unit
+    -> port:int -> dbpath:string -> path_in_url: string -> unit -> unit